Drop RECONFIGURE events from glimagesink
These are sent every time our GTK window is resized. The window is always
1024x768 on creation (hard coded in GStreamer). In windowed mode we then
resize it to match the input. In full screen mode further intermediate
sizes are set by the window manager until it's actually full screen.

Each resize sends a RECONFIGURE which forces reallocation of most
ION backed buffers. Since some are in flight we can run out of CMA
memory during the process for higher resolution frames.

This wasn't an issue in older GStreamer because the glmixer in
plugins-bad ate the events. glmixer in plugins-base is more clever and
can forward the events properly but this isn't needed in our use case.
We always use the same source resolution in the pipeline regardless of
sink size, and rely on the sink for scaling.

The RECONFIGURE can thus safely be dropped to avoid OOM.

Change-Id: I5dd8f8849d27e085e338e7219fc51963095e82f7